+++ dom/src/util.ts
@@ -4,8 +4,8 @@ import {
   TempOps,
   TempReal,
   TempSubOps,
-} from "better-react";
-import { useAttrEffect } from "better-react-helper";
+} from 'better-react'
+import { useAttrEffect } from 'better-react-helper'
 import {
   StoreRef,
   alawaysFalse,
@@ -13,96 +13,97 @@ import {
   quote,
   storeRef,
   diffMove,
-} from "wy-helper";
-import { renderChildrenOperate } from "wy-dom-helper";
-export type ContentEditable = boolean | "inherit" | "plaintext-only";
+} from 'wy-helper'
+import { renderChildrenOperate } from 'wy-dom-helper'
+export type ContentEditable = boolean | 'inherit' | 'plaintext-only'
 
 export function useContentEditable(
   node: ElementContentEditable,
-  contentEditable?: boolean | "inherit" | "plaintext-only",
+  contentEditable?: boolean | 'inherit' | 'plaintext-only'
 ) {
   useAttrEffect(() => {
-    node.contentEditable = contentEditable + "" || "true";
-  }, [node, contentEditable]);
+    node.contentEditable = contentEditable + '' || 'true'
+  }, [node, contentEditable])
 }
 
+const op = diffMove(renderChildrenOperate)
 export function nodeAppendChild(
   pNode: Node,
   list: ListCreater,
-  cache: StoreRef<readonly Node[]>,
+  cache: StoreRef<readonly Node[]>
 ) {
-  const lastChildren = cache.get();
-  const newChildren: Node[] = [];
-  addChildren(list, newChildren);
-  diffMove(renderChildrenOperate, pNode, lastChildren, newChildren);
-  cache.set(newChildren);
+  const lastChildren = cache.get()
+  const newChildren: Node[] = []
+  addChildren(list, newChildren)
+  op.move(pNode, lastChildren, newChildren)
+  cache.set(newChildren)
 }
 
 function addChild(row: Node | TempSubOps<ListCreater>, newChildren: Node[]) {
   if (row instanceof Node) {
-    newChildren.push(row);
+    newChildren.push(row)
   } else if (row instanceof TempSubOps) {
-    addChildren(row.data, newChildren);
+    addChildren(row.data, newChildren)
   } else {
-    console.error("不知道是什么类型,无法加入", row);
+    console.error('不知道是什么类型,无法加入', row)
   }
 }
 function addChildren(list: ListCreater, newChildren: Node[]) {
   for (let i = 0; i < list.list.length; i++) {
-    const row = list.list[i];
-    addChild(row, newChildren);
+    const row = list.list[i]
+    addChild(row, newChildren)
   }
 }
 
 export function createNodeTempOps(
   pel: Node,
-  createChangeAtom: CreateChangeAtom<any>,
+  createChangeAtom: CreateChangeAtom<any>
 ) {
-  const cache = storeRef(emptyArray);
-  const addEffect = createChangeAtom(false, alawaysFalse);
+  const cache = storeRef(emptyArray)
+  const addEffect = createChangeAtom(false, alawaysFalse)
   const root = new TempOps<ListCreater>(
     () => new ListCreater(),
     () => {
       if (!addEffect.get()) {
-        addEffect.set(true);
+        addEffect.set(true)
         hookEnvModel().updateEffect(-0.5, () => {
-          nodeAppendChild(pel, root.data, cache);
-        });
+          nodeAppendChild(pel, root.data, cache)
+        })
       }
-    },
-  );
-  return root;
+    }
+  )
+  return root
 }
 
 export class ListCreater implements TempReal {
-  list: (Node | TempSubOps<ListCreater>)[] = [];
+  list: (Node | TempSubOps<ListCreater>)[] = []
   reset(): void {
-    this.list.length = 0;
+    this.list.length = 0
   }
   add(...vs: any[]): void {
     for (let i = 0; i < vs.length; i++) {
-      const v = vs[i];
-      this.list.push(v);
+      const v = vs[i]
+      this.list.push(v)
     }
   }
 }
 
-const freeze = Object.freeze ? Object.freeze.bind(Object) : quote;
-export type TOrQuote<T extends {}> = T | ((v: T) => T | void);
+const freeze = Object.freeze ? Object.freeze.bind(Object) : quote
+export type TOrQuote<T extends {}> = T | ((v: T) => T | void)
 export function lazyOrInit<T extends {}>(v: TOrQuote<T>) {
-  let o: any;
-  if (typeof v == "function") {
+  let o: any
+  if (typeof v == 'function') {
     const obj = {
       style: {},
-    };
-    o = (v as any)(obj) || obj;
-    o.style = freeze(o.style);
-    o = freeze(o);
+    }
+    o = (v as any)(obj) || obj
+    o.style = freeze(o.style)
+    o = freeze(o)
   } else {
-    o = v;
+    o = v
     if (o) {
-      o = freeze(o);
+      o = freeze(o)
     }
   }
-  return o;
+  return o
 }