+++ core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # better-react
 
+## 2.0.0
+
+### Minor Changes
+
+- little change for performance
+
+### Patch Changes
+
+- Updated dependencies []:
+  - wy-helper@1.1.0
+
 ## 1.3.2
 
 ### Patch Changes