+++ demo/CHANGELOG.md
@@ -1,5 +1,17 @@
 # better-react-demo
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - better-react@2.0.0
+  - wy-dom-helper@2.0.0
+  - wy-helper@1.1.0
+  - better-react-dom@3.0.0
+  - better-react-dom-helper@3.0.0
+  - better-react-helper@3.0.0
+
 ## 0.0.4
 
 ### Patch Changes